requirements.txt
mcp>=1.0.0
httpx>=0.25.0
asyncio
typing-extensions>=4.5.0
PyYAML>=6.0